refactor(s1): move startDaemon to dkg-core/daemon-lifecycle
S1 step 2 of execution-plan.md §3.S1 (issue #386 — Hermes setup parity).
Behavior-unchanged extraction; `startDaemon` and its private helpers
(`hasLocalRepoForCli`, `blueGreenMigrationMayRunDuringStart`,
`daemonStartSpawnOptions`, `isProcessRunning`) move from adapter-openclaw
to @origintrail-official/dkg-core so adapter-hermes can reuse them in
S2 to satisfy issue #386 acceptance criterion 3 ("--no-start truly means
do not start the DKG daemon").

- Added: packages/core/src/daemon-lifecycle.ts
- Added: export via packages/core/src/index.ts
- adapter-openclaw/src/setup.ts now thin-re-exports startDaemon from
  dkg-core (preserves the adapter's public surface)
- Removed unused imports from adapter-openclaw/src/setup.ts: spawnSync,
  SpawnSyncOptions, lstatSync, blueGreenSlotReady, findPackageRepoDir,
  resolveDkgCli; kept `realpathSync` (still used by step 8b migration
  cleanup) and `sleep` (still used by readWalletsWithRetry — extracted
  in S1.3)
- adapter-openclaw/test/setup-start-daemon.test.ts mock target updated
  from '../src/resolve-dkg-cli.js' to dkg-core's
  resolve-dkg-cli.js dist path (vitest needs to intercept the real
  cross-module call path inside daemon-lifecycle.ts)

The `[setup] ...` console.log prefix is preserved verbatim in the new
core module so user-visible output is unchanged.

Test results vs s1-baseline.md (post-extraction):
- @origintrail-official/dkg-core: 33 files, 550 tests — green
- @origintrail-official/dkg-adapter-openclaw: 22 files, 974 tests —
  same 2 pre-existing failures as baseline; zero new failures
- @origintrail-official/dkg-adapter-hermes: 1 file, 60 tests — green
- @origintrail-official/dkg-node-ui: 20 files — green
- @origintrail-official/dkg (CLI): curated subset gate per
  s1-baseline.md (Hardhat-dependent CLI tests deferred to release-readiness)
