+++ packages/core/src/index.ts
@@ -28,6 +28,7 @@ export {
 export { requestFaucetFunding, type FaucetResult } from './faucet.js';
 export { resolveCliPackageDir } from './resolve-cli-package-dir.js';
 export { resolveDkgCli, type ResolvedDkgCli } from './resolve-dkg-cli.js';
+export { startDaemon } from './daemon-lifecycle.js';
 export {
   assertSafeIri,
   isSafeIri,
